+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Although Moulti's Python packages, modules and functions are obviously available
 
 - CLI command `moulti question {add,update,get-answer,delete}`
 - Ansible callback plugin: support `prompt_until()` and the `pause` module
+- quit dialog: buttons now feature keyboard accelerators
 
 ### Changed